# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.4.0
  description: Oogly Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages (paginated, newest first)
      parameters:
        - in: query
          name: before
          schema:
            type: integer
          description: Cursor for pagination (message id)
        - in: query
          name: limit
          schema:
            type: integer
          description: Max number of messages to return
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request (e.g. filtered content)
        "403":
          description: Forbidden (banned or muted)
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (admin or own)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Reactions
  /messages/{id}/reactions:
    post:
      operationId: addReaction
      tags: [reactions]
      summary: Add a reaction to a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "201":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
  /messages/{id}/reactions/{emoji}:
    delete:
      operationId: removeReaction
      tags: [reactions]
      summary: Remove a reaction from a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: emoji
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /messages/{id}/reactions/summary:
    get:
      operationId: getReactionSummary
      tags: [reactions]
      summary: Get reaction counts per emoji for a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReactionSummary"

  # Users
...
  /mutes:
    get:
      operationId: listMutes
      tags: [mutes]
      summary: List all mutes (admin)
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: Mutes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mute"

    post:
      operationId: muteUser
      tags: [mutes]
      summary: Mute a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuteInput"
      responses:
        "201":
          description: Muted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mute"
  /mutes/{id}:
    delete:
      operationId: unmuteUser
      tags: [mutes]
      summary: Remove a mute (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unmuted

  # Appeals
...
  /filter/check:
    post:
      operationId: checkFilter
      tags: [admin]
      summary: Test a string against the swear filter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterCheckInput"
      responses:
        "200":
          description: Filter result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterCheckResult"
components:
  schemas:
...
    Message:
      type: object
      required: [id, content, authorId, authorUsername, deleted, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        deleted:
          type: boolean
        replyToId:
          type: ["integer", "null"]
        replyToContent:
          type: ["string", "null"]
        replyToUsername:
          type: ["string", "null"]
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionSummary"
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        replyToId:
          type: integer

    # Reactions
    Reaction:
      type: object
      required: [id, messageId, userId, emoji, createdAt]
      properties:
        id:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        emoji:
          type: string
        createdAt:
          type: string
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
    ReactionSummary:
      type: object
      required: [emoji, count, userReacted]
      properties:
        emoji:
          type: string
        count:
          type: integer
        userReacted:
          type: boolean

    # Bans
...
    Mute:
      type: object
      required: [id, userId, reason, isPermanent, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        mutedById:
          type: ["integer", "null"]
        mutedByUsername:
          type: ["string", "null"]
        reason:
          type: string
        isPermanent:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    MuteInput:
      type: object
      required: [userId, reason]
      properties:
        userId:
          type: integer
        reason:
          type: string
          minLength: 1
        isPermanent:
          type: boolean
        expiresAt:
          type: ["string", "null"]

    # Appeals
...
    FilterCheckInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
    FilterCheckResult:
      type: object
      required: [flagged, matches, cleaned]
      properties:
        flagged:
          type: boolean
        matches:
          type: array
          items:
            type: string
        cleaned:
          type: string