# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oogly Chat API specification
 * OpenAPI spec version: 1.4.0
 */

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
  replyToId?: number;
}