# lib/api-zod/src/generated/types/mute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Oogly Chat API specification
 * OpenAPI spec version: 1.4.0
 */

export interface Mute {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  mutedById?: number | null;
  /** @nullable */
  mutedByUsername?: string | null;
  reason: string;
  isPermanent: boolean;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}